AVF: provide vsock_socket perms to vendor_avf_qcvm_manager
vendor_avf_qcvm_manager as a vsock server and it binds to CID_HOST and port, then waiting for client connection. we need to provide vsock_socket permissions to vendor_avf_qcvm_manager. we bind vendor_avf_qcvm_manager to unconstrained_vsock_violators attribute because it is an exemption to vsock neverallow in system side.

Change-Id: I4b793af7b693eff43a61648f083d898a0a1edb27