AVF: Update AVF manager SEPolicy
Currently Mink hub daemon requires vendor_mink_trace_prop access to 
read the trace properties when accessing the VM's. This change is 
to provide access to the trace props during Mink communication.

The early virt mgr has a capability to write kernel messages. 
This change is providing the ability to write the kernel message 
using the capability provided.

Change-Id: I498a97fa703508ad9cd2e5fa4d96ddef29adc600